requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.16.0
chardet==4.0.0
cryptography==42.0.2
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.61.1
django-menu==0.1.13
django-model-utils==4.4.0
djangorestframework==3.14.0
idna==2.10
oauthlib==3.2.2
Pillow==8.2.0
pycparser==2.21
PyJWT==2.8.0
PyPDF2==3.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.1
stripe==2.58.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.5
